[UI] Phase 5.a: Split filters/Filters.tsx into focused siblings (#18660)
Splits the 1258-line filters/Filters.tsx into focused sibling files,
mirroring the MesheryPatterns.tsx pattern. Entry point drops to ~440
measured lines; over_hard giant-files count drops by 1.

New sibling files in ui/components/filters/:
  Filters.styled.tsx     styled() primitives (CreateButton, BtnText, ...)
  Filters.types.ts       shared TypeView / props / payload types
  Filters.constants.ts   ACTION_TYPES and column-visibility tuple
  TooltipIcon.tsx        the small CustomTooltip-wrapped IconButton
  YAMLEditor.tsx         the inline CodeMirror dialog for filter config
  ImportModal.tsx        the RJSF Import Design modal
  PublishModal.tsx       the RJSF catalog-publish modal
  Filters.columns.tsx    buildFiltersColumns() factory
  Filters.tableOptions.tsx  buildFiltersTableOptions() factory
  Filters.fileActions.ts    create*() factories for filter actions

Entry-point export and prop signature are unchanged. ESLint and tsc
clean; legacyMaxLineOffenders no longer lists Filters.tsx.

Refs #18660
Refs #18654
